pyproject.toml
[tool.poetry]
name = "amastay"
version = "0.2.0"
description = "Amastay AI"
authors = ["Brian Perin <brian@amastay.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
boto3 = "^1.35.29"
huggingface-hub = "^0.25.1"
httpx = "^0.27.2"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
requests-aws4auth = "^1.3.1"
pytest = "^8.3.3"
selenium = "^4.25.0"
furl = "^2.1.3"
PyJWT = "^2.9.0"
uuid = "^1.30"
python-dotenv = "^1.0.1"
supabase = "^2.8.0"
setuptools = "^75.1.0"
sagemaker = "2.232.1"
pydantic = ">=2.5.3,<2.9.0"
geopy = "^2.4.1"
stripe = "^11.1.1"
openapi-generator-cli = "^4.3.1"
undetected-chromedriver = "^3.5.5"
apscheduler = "^3.10.4"
phonenumbers = "^8.13.49"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
huggingface = "^0.0.1"
email-validator = "^2.2.0"
fastapi-pagination = "^0.12.32"

[tool.poetry.group.dev.dependencies]
pyright = "^1.1.382.post1"

[tool.poetry.scripts]
start = "uvicorn app:app --host=0.0.0.0 --port=80"
dev = "uvicorn app:app --reload --port 5001"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 400
target-version = ['py39']
include = '\.pyi?$'
